legal-doc-qa-rag/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.6
chromadb>=0.4.22
sentence-transformers>=2.2.2
tiktoken>=0.5.0
numpy>=1.24.0
